[module12-task1] перенёс создание генератора уникальных случайных чисел в utils
