+++ js/utils/index.js
@@ -20,4 +20,34 @@ function getRandArrayElem(anArray) {
   return anArray[getRandInt(0, anArray.length - 1)];
 }
 
-export { getRandArrayElem, getRandInt };
+/**
+  * @description Возвращает функцию, которая генерирует случайное не повторяющееся целове
+  *             число из диапазона. Если все числа диапазона были возвращены,
+  * @param {number} from начало отрезка (включительно)
+  * @param {number} to конец отрезка (включительно)
+  * @returns {number} уникальное случайное число из отрезка
+  */
+function createUniqueRandIntGenerator(from, to) {
+  const _from = Math.min(Math.floor(from), Math.floor(to));
+  const _to = Math.max(Math.floor(from), Math.floor(to));
+
+  const prevValues = new Set();
+  return () => {
+    //console.log(prevValues.size);
+    if (prevValues.size >= _to - _from + 1) {
+      //console.error('Все значения закончились');
+      return null;
+    }
+    let newVal = getRandInt(_from, _to);
+    while (prevValues.has(newVal)) {
+      newVal = getRandInt(_from, _to);
+    }
+
+    prevValues.add(newVal);
+
+    return newVal;
+  };
+}
+
+
+export { getRandArrayElem, getRandInt, createUniqueRandIntGenerator };
